move signing to be release only
refs #1

this helps fix the initial "build fail due to missing key" if someone
just checks out the project loads it up, or tries to build it in debug
mode. since releases are really the only things we should be signing it
was quick to move it in there.

long term, we still want folks being able to build release builds
locally too, without a key/generating one, and only sign the real
releases we distribute to folks, but this gets me unblocked to start
porting the build from MSBuild (something windows only), to the newer
dotnet build stuff that is cross platform